requirements.txt
GPUtil==1.4.0
ipython==7.20.0
matplotlib==3.3.4
numpy==1.20.0
opencv_python==4.5.2.54
pandas==1.2.2
pre-commit==2.13.0
pylint==2.8.2
pytest==6.2.4
PyYAML==5.4.1
scipy==1.4.1
seaborn==0.11.1
simpleitk==2.0.2
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-profile==2.4.0
tensorflow-gpu==2.3.0
tqdm==4.56.1
